refactor: move task card to ui components
